mavlink_commands: drop _state
In my opinion we dont really need the _state variable because we can
derive what is going on from set timeouts and what is in the work queue.
All that we actually need is to know if we already sent an item which
has been added as a flag to each work item.